+++ customize/keymaps.py
@@ -104,6 +104,8 @@ class FOCUS_SELECTED_WITH_F(keymap_utils.KeymapGroup):
         self.km = kc.keymaps.new('Outliner', space_type='OUTLINER')
         self.add('clip.view_selected', key, 'PRESS')
 
+        # self.add('outliner.armored_cycle', 'C', 'PRESS')
+
         self.enabled_message()
 
     
@@ -151,13 +153,16 @@ class SCULPTING_SETUP(keymap_utils.KeymapGroup):
     def register(self):
         self.km = kc.keymaps.new(name='Sculpt')
 
-        self.add('object.transfer_mode', 'D', 'PRESS')
-        # self.add('view3d.view_center_pick', 'F', 'PRESS')
-        self.add('view3d.armored_focus', 'F', 'PRESS')
-        self.add('view3d.view_all', 'F', 'PRESS', shift=True)
-        self.add('view3d.armored_silhouette_mode', 'V', 'PRESS')
-        # self.add('view3d.view_center_pick', 'F', 'PRESS', alt=True)
+        # self.add('view3d.armored_focus', 'C', 'PRESS')
+        self.add('view3d.armored_silhouette', 'V', 'PRESS')
         self.add('view3d.armored_subdivide', 'D', 'PRESS', ctrl=True)
+        self.add('sculpt.armored_remesh', 'R', 'PRESS', ctrl=True)
+
+        # self.add('view3d.view_all', 'F', 'PRESS', shift=True)
+
+        # self.add('sculpt.armored_scale_unmasked', 'S', 'PRESS', alt=True)
+        self.add('transform.translate', 'G', 'PRESS')
+
 
         # Invert brush stroke (set to ALT instead of CTRL)
         self.add('sculpt.brush_stroke', 'LEFTMOUSE', 'PRESS', alt=True); self.prop('mode', 'INVERT')
@@ -182,7 +187,7 @@ class SCULPTING_SETUP(keymap_utils.KeymapGroup):
         self.add('wm.context_toggle',  'T', 'PRESS', ctrl=True);  self.prop('data_path', 'space_data.overlay.show_fade_inactive')
 
         # BRUSH SIZE
-        self.add('wm.radial_control',  'S', 'PRESS');
+        self.add('wm.radial_control',  'F', 'PRESS');
         self.prop('data_path_primary',   'tool_settings.sculpt.brush.size.size')
         self.prop('data_path_secondary', 'tool_settings.unified_paint_settings.size')
         self.prop('use_secondary',       'tool_settings.unified_paint_settings.use_unified_size')
@@ -197,7 +202,7 @@ class SCULPTING_SETUP(keymap_utils.KeymapGroup):
         self.prop('release_confirm', True)  # Only setting that was changed.
 
         # BRUSH STRENGTH
-        self.add('wm.radial_control',  'S', 'PRESS', shift=True);
+        self.add('wm.radial_control',  'F', 'PRESS', shift=True);
         self.prop('data_path_primary',   'tool_settings.sculpt.brush.size.strength')
         self.prop('data_path_secondary', 'tool_settings.unified_paint_settings.strength')
         self.prop('use_secondary',       'tool_settings.unified_paint_settings.use_unified_strength')